refactor: NewTicketPage 위치 이동
